+++ artifex/models/classification/multi_label_classification/multi_label_classification_model.py
@@ -32,7 +32,8 @@ class MultiLabelClassificationModel(BaseModel):
     
     def __init__(
         self, synthex: Synthex,
-        tokenizer_max_length: int = config.DEFAULT_TOKENIZER_MAX_LENGTH
+        tokenizer_max_length: int = config.DEFAULT_TOKENIZER_MAX_LENGTH,
+        base_model_name: Optional[str] = None
     ):
         super().__init__(synthex)
         self._synthetic_data_schema_val: JobOutputSchemaDefinition = {
@@ -40,7 +41,7 @@ class MultiLabelClassificationModel(BaseModel):
             "labels": {"type": "string"},
         }
         self._token_keys_val: list[str] = ["text"]
-        self._base_model_name_val: str = config.CLASSIFICATION_HF_BASE_MODEL
+        self._base_model_name_val: str = base_model_name or config.CLASSIFICATION_HF_BASE_MODEL
         self._tokenizer_max_length_val: int = tokenizer_max_length
         self._system_data_gen_instr_val: list[str] = [
             "The 'text' field should contain text that belongs to the following domain(s): {domain}.",